Set BUILD_LIBRARY_FOR_DISTRIBUTION to YES.
This should allow the binary generated by compiling the framework with any Swift 5.1+ compiler to be used with future Swift compilers.  For example, after building the library on Xcode 11.5, it will still be usable on Xcode 12 rather than producing a "Module compiled with Swift 5.2.4 cannot be imported by the Swift 5.3 compiler" error.

In order to compile the library with this enabled, any dynamic properties must be decorated @objc.